requirements.txt
flet==0.12.0
fastapi==0.95.0
uvicorn==0.18.2
pydantic==1.10.0
pymongo==4.3.3
python-dotenv==0.19.2
requests==2.26.0
matplotlib
python-jose